requirements.txt
python-decouple==3.4
bluepy==1.3.0
requests